feat: finished Shortcut
