+++ OceanDesignSystem/Controllers/Components/ShortcutViewController.swift
@@ -12,64 +12,6 @@ import OceanTokens
 import OceanComponents
 
 final public class ShortcutViewController : UIViewController {
-
-    let subtitleTextTiny = "Lorem ipsum"
-    let subtitleTextSmall = "Lorem ipsum dolor sit amet, consectetur."
-    let subtitleTextMedium = "Lorem ipsum dolor sit amet, consectetur. Lorem ipsum dolor sit amet, consectetur."
-
-//    private lazy var shortcut1: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .tiny
-//        view.orientation = .vertical
-//
-//        return view
-//    }()
-//
-//    private lazy var shortcut2: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .tiny
-//        view.orientation = .horizontal
-//
-//        return view
-//    }()
-//
-//    private lazy var shortcut3: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .small
-//        view.orientation = .vertical
-//
-//        return view
-//    }()
-//
-//    private lazy var shortcut4: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .small
-//        view.orientation = .horizontal
-//
-//        return view
-//    }()
-//
-//    private lazy var shortcut5: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .medium
-//        view.orientation = .vertical
-//
-//        return view
-//    }()
-//
-//    private lazy var shortcut6: Ocean.NewShortcut = {
-//        let view = Ocean.NewShortcut()
-//        view.translatesAutoresizingMaskIntoConstraints = false
-//        view.size = .medium
-//        view.orientation = .horizontal
-//
-//        return view
-//    }()
     
     public override func viewDidLoad() {
         self.view.backgroundColor = .white
@@ -93,68 +35,6 @@ final public class ShortcutViewController : UIViewController {
             .fill(to: scrollView)
             .width(to: view, constant: -(Ocean.size.spacingInsetSm * 2))
             .make()
-
-//        addSection(stackView: contentStack, text: "Tiny - Vertical")
-//        contentStack.addArrangedSubview(shortcut1)
-//        addSection(stackView: contentStack, text: "Tiny - Horizontal")
-//        contentStack.addArrangedSubview(shortcut2)
-//        addSection(stackView: contentStack, text: "Small - Vertical")
-//        contentStack.addArrangedSubview(shortcut3)
-//        addSection(stackView: contentStack, text: "Small - Horizontal")
-//        contentStack.addArrangedSubview(shortcut4)
-//        addSection(stackView: contentStack, text: "Medium - Vertical")
-//        contentStack.addArrangedSubview(shortcut5)
-//        addSection(stackView: contentStack, text: "Medium - Horizontal")
-//        contentStack.addArrangedSubview(shortcut6)
-
-//        addExample(badgeNumber: 1,
-//                   title: subtitleTextTiny,
-//                   subtitle: subtitleTextTiny,
-//                   blocked: false)
-//
-//        addExample(badgeNumber: 1,
-//                   title: subtitleTextTiny,
-//                   subtitle: subtitleTextTiny,
-//                   blocked: false)
-//
-//        addExample(badgeNumber: 1,
-//                   title: subtitleTextSmall,
-//                   subtitle: subtitleTextSmall,
-//                   blocked: false)
-//
-//        addExample(badgeNumber: 1,
-//                   title: subtitleTextMedium,
-//                   subtitle: subtitleTextMedium,
-//                   blocked: false)
-//
-//        addExample(badgeNumber: 1,
-//                   title: subtitleTextMedium,
-//                   subtitle: subtitleTextMedium,
-//                   blocked: false)
-
-//        shortcut1.setData(with: examples, cols: 2)
-//        shortcut2.setData(with: examples, cols: 2)
-//        shortcut3.setData(with: examples, cols: 2)
-//        shortcut4.setData(with: examples, cols: 2)
-//        shortcut5.setData(with: examples, cols: 2)
-//        shortcut6.setData(with: examples, cols: 2)
-
-//        addSection(stackView: contentStack, text: "Tiny - Vertical")
-//        contentStack.addArrangedSubview(shortcut1)
-//        addSection(stackView: contentStack, text: "Tiny - Horizontal")
-//        contentStack.addArrangedSubview(shortcut2)
-//        addSection(stackView: contentStack, text: "Small - Vertical")
-//        contentStack.addArrangedSubview(shortcut3)
-//        addSection(stackView: contentStack, text: "Medium - Vertical")
-//        contentStack.addArrangedSubview(shortcut5)
-//        addSection(stackView: contentStack, text: "Medium - Horizontal")
-//        contentStack.addArrangedSubview(shortcut6)
-//        addSection(stackView: contentStack, text: "Carousel")
-//        contentStack.addArrangedSubview(shortcut7)
-//
-//        shortcut1.onTouch = { index in
-//            print(index)
-//        }
 
         addSection(stackView: contentStack, text: "Tiny - Vertical")
         addExample(stack: contentStack,
@@ -183,31 +63,6 @@ final public class ShortcutViewController : UIViewController {
                             badgeNumber: nil,
                             badgeStatus: .neutral,
                             title: "Label",
-                            subtitle: "",
-                            blocked: false),
-        Ocean.ShortcutModel(image: Ocean.icon.documentOutline,
-                            badgeNumber: 0,
-                            badgeStatus: .neutral,
-                            title: "Label",
-                            subtitle: "",
-                            blocked: false),
-        Ocean.ShortcutModel(image: Ocean.icon.documentOutline,
-                            badgeNumber: 1,
-                            badgeStatus: .highlight,
-                            title: "Label",
-                            subtitle: "",
-                            blocked: false),
-        Ocean.ShortcutModel(image: Ocean.icon.documentOutline,
-                            badgeNumber: nil,
-                            badgeStatus: .highlight,
-                            title: "Label",
-                            subtitle: "",
-                            blocked: true),
-
-        Ocean.ShortcutModel(image: Ocean.icon.documentOutline,
-                            badgeNumber: nil,
-                            badgeStatus: .neutral,
-                            title: "Label",
                             subtitle: "Lorem ipsum dolor sit amet, consectetur. Lorem ipsum dolor sit amet, consectetur.",
                             blocked: false),
         Ocean.ShortcutModel(image: Ocean.icon.documentOutline,
@@ -231,18 +86,10 @@ final public class ShortcutViewController : UIViewController {
     ]
 
     public func addExample(stack: UIStackView,
-                           orientation: Ocean.NewShortcut.Orientation,
-                           size: Ocean.NewShortcut.Size) {
-//        let model: Ocean.ShortcutModel = .init(image: Ocean.icon.documentOutline,
-//                                               badgeNumber: badgeNumber,
-//                                               badgeStatus: (badgeNumber ?? 0) > 0 ? .highlight : .neutral,
-//                                               title: title,
-//                                               subtitle: subtitle,
-//                                               blocked: blocked)
-//
-//        examples.append(model)
+                           orientation: Ocean.Shortcut.Orientation,
+                           size: Ocean.Shortcut.Size) {
 
-        let view = Ocean.NewShortcut()
+        let view = Ocean.Shortcut()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.orientation = orientation
         view.size = size
@@ -253,62 +100,7 @@ final public class ShortcutViewController : UIViewController {
 
         stack.addArrangedSubview(view)
     }
-    
-//    public override func viewDidAppear(_ animated: Bool) {
-//        super.viewDidAppear(animated)
-//
-//        // Tiny - Vertical
-//        shortcut1.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label")
-//        ], quantityPage: 2)
-//
-//        // Tiny - Horizontal
-//        shortcut2.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label")
-//        ], quantityPage: 2)
-//
-//        // Small - Vertical
-//        shortcut3.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label", subtitle: subtitleText)
-//        ], quantityPage: 2)
-//
-//        // Medium - Vertical
-//        shortcut5.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label", subtitle: subtitleText)
-//        ], quantityPage: 2)
-//
-//        // Medium - Horizontal
-//        shortcut6.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label", subtitle: subtitleText),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label", subtitle: subtitleText)
-//        ], quantityPage: 2)
-//
-//        // Carousel
-//        shortcut7.addData(with: [
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 0, badgeStatus: .neutral, title: "Label"),
-//            Ocean.ShortcutModel(image: Ocean.icon.documentOutline!, badgeNumber: 100, badgeStatus: .highlight, title: "Label")
-//        ], quantityPage: 2)
-//    }
-//
+
     private func addSection(stackView: Ocean.StackView, text: String, includeSpacer: Bool = true) {
         let label = Ocean.Typography.description { label in
             label.translatesAutoresizingMaskIntoConstraints = false
